% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLASSOAdaptive.R
\name{BayesGLASSOAdaptive}
\alias{BayesGLASSOAdaptive}
\title{Adaptive Bayesian Graphical LASSO.}
\usage{
BayesGLASSOAdaptive(data, lambda_shape = 0.01, lambda_rate = 1e-06,
  nBurnin = 10000, nIter = 10000, verbose = TRUE, lambdaDiag = 1)
}
\arguments{
\item{data}{A data matrix with rows representinh participants and columns
representing variables/nodes.}

\item{lambda_shape}{The rate parameter for the hyperprior of lambda, the parameter
that determines the amount of shirnkage.}

\item{lambda_rate}{The scale parameter for the hyperprior of lambda, the parameter
that determines the amount of shirnkage.}

\item{nIter}{The number of iterations for the Gibbs-sampler.}

\item{verbose}{If 'TRUE'displays a progress bar.}

\item{lambdaDiag}{The shrinkage parameter value to be used for diagonal elements.}

\item{nBurning}{The number of burn-in iterations for the Gibbs-sampler.}
}
\value{
The function returns the following output:
\item{Omega}{A dataframe containing the estimated precision matrix for each itteration of the Gibbs-sampler.}
\item{pcor}{A dataframe containing the estimated partial correlation matrix for each itteration of the Gibbs-sampler.}
\item{lambda_sq}{A dataframe containing the estimates for the local hyper parameter lambda for each itteration of the Gibbs-sampler.}
\item{tau_sq}{A dataframe containing the estimates for the global hyper parameter tau for each itteration of the Gibbs-sampler.}
\item{Strength}{A dataframe containing the estimated strength centrality for each node for each itteration of the Gibbs-sampler.}
\item{Closeness}{A dataframe containing the estimated closeness centrality for each node for each itteration of the Gibbs-sampler.}
\item{Betweenness}{A dataframe containing the estimated betweenness centrality for each node for each itteration of the Gibbs-sampler.}
\item{optwi}{The point estimate for the precision matrix obtained by taking the mode of the posterior distribution.}
\item{optpcor}{The point estimate for the partial correlation matrix obtained by taking the mode of the posterior distribution.}
\item{CredInt}{The 95\% Credibility interval for the elements of the precision matrix.}
\item{optStrength}{The point estimate for strength centrality of each node based on the point estimate of the partial correlation matrix.}
\item{CIStrength}{The 95\% Credibility Interval for the strength centrality of each node.}
\item{optCloseness}{The point estimate for closeness centrality of each node based on the point estimate of the partial correlation matrix.}
\item{CICloseness}{The 95\% Credibility Interval for the closeness centrality of each node.}
\item{optBetween}{The point estimate for betweenness centrality of each node based on the point estimate of the partial correlation matrix.}
\item{CIBetween}{The 95\% Credibility Interval for the betweenness centrality of each node.}

The following output is only returned in case of missing data.

\item{Missing}{Indicator for which observations where missing.}
\item{ImputedValues}{The imputed values for each missing datapoint for each itteration of the Gibbs-sampler.}
\item{CompleteData}{The original dataset with missing values replaced by the mean imputed value for each missing data point.}
\item{OriginalData}{The original dataset with missing values.}
}
\description{
\code{BayesGLASSOAdaptive} Provides a regularized precision matrix estimate using
a Bayesian GLASSO, in which different shrinkage parameter values are used for
each off-diagonal element.
}
