% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRegularizedWishart.R
\name{BayesRegularizedWishart}
\alias{BayesRegularizedWishart}
\title{Regularized Wishart Estimator.}
\usage{
BayesRegularizedWishart(data, adaptive = FALSE, adfact = 2)
}
\arguments{
\item{data}{A data matrix with rows representinh participants and columns
representing variables/nodes.}

\item{adaptive}{If 'TRUE' chose the amount of shrinkage based on the ration of n to p.}

\item{adfact}{the ration between the two shrinkage hyper-parameters used.}
}
\value{
Returns dataframes for the precision matrix, partial
correlations, and an overview of the edges fixed to 0 with estimates from each
iteration of the Gibbs-sampler.
}
\description{
\code{BayesRegularizedWishart} Provides a ridge type regularized precision matrix estimate using
an eigenvalue and eigenvector decomposition of the sample covariance matrix.
}
