% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesWishart.R
\name{BayesWishart}
\alias{BayesWishart}
\title{Wishart Estimator.}
\usage{
BayesWishart(data, mu0 = 0)
}
\arguments{
\item{data}{A data matrix with rows representinh participants and columns
representing variables/nodes.}
}
\value{
Returns dataframes for the precision matrix, partial
correlations, and the variability is the estimates of the precision matrix entries
with estimates from each iteration of the Gibbs-sampler.
}
\description{
\code{BayesWishart} Provides non-regularized precision matrix estimate.
}
